refactor(security): add JwtService, move authenticate to AuthService from UserService, fix user update on actions
