+++ security-service/src/main/java/com/ilogos/security/user/UserService.java
@@ -5,14 +5,10 @@ import java.util.Collection;
 import java.util.Optional;
 
 import org.springframework.http.HttpStatus;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.DisabledException;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.core.AuthenticationException;
 import org.springframework.stereotype.Service;
 
-import com.ilogos.security.auth.AuthService;
 import com.ilogos.security.exception.ExceptionWithStatus;
+import com.ilogos.security.jwt.JwtService;
 import com.ilogos.security.user.UserController.UpdateUserRequest;
 import com.ilogos.security.user.emailHistory.EmailHistory;
 import com.ilogos.security.user.emailHistory.EmailHistoryRepository;
@@ -30,12 +26,11 @@ import lombok.extern.slf4j.Slf4j;
 @RequiredArgsConstructor
 public class UserService {
 
-    private final AuthService authService;
+    private final JwtService jwtService;
 
     private final UserRepository userRepository;
     private final UsernameHistoryRepository usernameHistoryRepository;
     private final EmailHistoryRepository emailHistoryRepository;
-    private final AuthenticationManager authenticationManager;
 
     public record TokensData(String accessToken, String refreshToken) {
     }
@@ -64,33 +59,23 @@ public class UserService {
         return user;
     }
 
-    public Optional<TokensData> authenticate(String usernameOrEmail, String password) {
-        log.info("User auth: {}", usernameOrEmail);
+    public record UserWithTokens(User user, TokensData tokens) {
+    }
+
+    public Optional<UserWithTokens> generateTokens(String usernameOrEmail, String password) {
         return userRepository.findByEmailOrUsername(usernameOrEmail, usernameOrEmail).map(user -> {
-            try {
-                authenticationManager.authenticate(
-                        new UsernamePasswordAuthenticationToken(usernameOrEmail, password));
+            var tokens = generateTokens(user, Optional.empty());
 
-                var tokens = generateTokens(user, Optional.empty());
-                user.setLastTokenIssuedAt(authService.getTokenInfo(tokens.accessToken), true);
+            user.setLastTokenIssuedAt(jwtService.getTokenInfo(tokens.accessToken), true);
+            update(user);
 
-                log.info("Auth success: {}", usernameOrEmail);
-
-                return tokens;
-            } catch (DisabledException ex) {
-                log.info("Auth error (disabled): {}", usernameOrEmail);
-                throw new ExceptionWithStatus(HttpStatus.UNAUTHORIZED, ex);
-            } catch (AuthenticationException ex) {
-                log.info("Auth error: {}", usernameOrEmail);
-                updateFailedAttempts(user);
-                throw new ExceptionWithStatus(HttpStatus.UNAUTHORIZED, ex);
-            }
+            return new UserWithTokens(user, tokens);
         });
     }
 
     private TokensData generateTokens(User user, Optional<String> optionalRefreshToken) {
-        String accessToken = authService.generateToken(user, true);
-        String refreshToken = optionalRefreshToken.orElseGet(() -> authService.generateToken(user, false));
+        String accessToken = jwtService.generateToken(user, true);
+        String refreshToken = optionalRefreshToken.orElseGet(() -> jwtService.generateToken(user, false));
 
         return new TokensData(accessToken, refreshToken);
     }
@@ -102,7 +87,8 @@ public class UserService {
             String username = tokenInfo.getUsername();
 
             var tokens = generateTokens(user, Optional.of(tokenInfo.getToken()));
-            user.setLastTokenIssuedAt(authService.getTokenInfo(tokens.accessToken), false);
+            user.setLastTokenIssuedAt(jwtService.getTokenInfo(tokens.accessToken), false);
+            update(user);
 
             log.info("Token refresh success: {}", username);
 
@@ -135,7 +121,7 @@ public class UserService {
         return update(user);
     }
 
-    public Optional<User> updateByAuth(TokenInfo tokenInfo, UpdateUserRequest request) {
+    public Optional<User> updateSelf(TokenInfo tokenInfo, UpdateUserRequest request) {
         Optional<User> user = userRepository.findById(tokenInfo.getId());
         return user.map(e -> {
             request.email().ifPresent(e::setEmail);
